rpmsg: disable hex dump traces
There are couple of print_hex_dump traces used in rpmsg code which
prints the actual byte messages being transferred between host and
the remote processors. These traces are quiet verbose and affects
performance, if the appropriate trace level is enabled. This patch
disables them by commenting them out.

They are intentionally not removed since they are core debug traces
and are useful when debugging complex IPC issues. It is easier to
enable them in code when needed rather than having to look up and
add the trace later.

Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>