drivers/rtc/rtc-pl031.c: restore ST variant functionality
e7e034e18a0ab ("drivers/rtc/rtc-pl031.c: fix the missing operation on
enable") accidentally broke the ST variants of PL031.

The bit that is being poked as "clockwatch" enable bit for the ST variants
does the work of bit 0 on this variant.  Bit 0 is used for a clock divider
on the ST variants, and setting it to 1 will affect timekeeping in a real
bad way.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Acked-by: Haojian Zhuang <haojian.zhuang@gmail.com>
Cc: Mian Yousaf KAUKAB <mian.yousaf.kaukab@stericsson.com>
Cc: Srinidhi Kasagar <srinidhi.kasagar@stericsson.com>
Cc: Alessandro Zummo <a.zummo@towertech.it>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>