printk: avoid softlockups in console_unlock()
A CPU can be caught in console_unlock() for a long time (tens of seconds
are reported by our customers) when other CPUs are using printk heavily
and serial console makes printing slow.  Despite serial console drivers
are calling touch_nmi_watchdog() this triggers softlockup warnings because
interrupts are effectively disabled for the whole time printing takes
place.  Thus IPIs cannot be processed and other CPUs get stuck spinning in
calls like smp_call_function_many().  Also RCU eventually starts reporting
lockups.

In my artifical testing I also managed to trigger a situation when disk
disappeared from the system apparently because commands to / from it could
not be delivered for long enough.  This is why just silencing watchdogs
isn't a reliable solution to the problem.

One part of fixing the issue is changing vprintk_emit() to call
console_unlock() with interrupts enabled (this isn't perfect as printk()
itself can be called with interrupts disabled but it improves the
situation in lots of cases).  Another part is limiting the time we spend
in console_unlock() printing loop to watchdog_thresh() / 4.  Then we
release console_sem and wait for watchdog_thresh() / 4 to give a chance to
other printk() users to get the semaphore and start printing.  If printk()
was called with interrupts enabled, it also gives CPU a chance to process
blocked interrupts.  Then we recheck if there's still anything to print,
try to grab console_sem again and if we succeed, we go on with printing.

Signed-off-by: Jan Kara <jack@suse.cz>
Cc: "Paul E. McKenney" <paulmck@us.ibm.com>
Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Frederic Weisbecker <fweisbec@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>