regulator: ti-avs-class0: use regulator name based on device instance
Don't use the same regulator descriptor for all regulator instance,
as there will be a name conflict. Instead pick up the device name and
create the descriptor.

As part of this change, save on multiple kzallocs by moving the
descriptor allocation as part of tiavs_class0_data allocation and drop
the need for an temp descriptor.

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Keerthy <j-keerthy@ti.com>