ARM: dts: DRA7: Add IPU timer capabilities
Add the ti,timer-ipu capability to a couple of DRA7xx DMTimers
(currently limited to the ones that have default interrupt mappings
into the IPU processors, they are the same set as those on OMAP4).
This enables the timers to be requested by the remoteproc code
using the timer capabilities.

NOTE: No distinction has been made between IPU1 and IPU2, both of
them have the same default interrupt mappings. The timers are
expected to be assigned statically between the two processors from
the common set. There are some more timers actually that are capable
of interrupting an IPU through a crossbar, but this is beyond the
current scope since the crossbar programming for a particular
processor is expected to be handled by the code running on that
processor.

Signed-off-by: Suman Anna <s-anna@ti.com>