iommu/omap: add support for configuring dsp iommus on DRA7xx
The DSP MMUs on DRA7xx SoC requires configuring an additional
MMU_CONFIG register present in the DSP_SYSTEM address space.
This setting dictates whether the DSP Core's MDMA and EDMA
traffic is routed through the respective MMU or not. Add the
support to the OMAP iommu driver so that the traffic is not
bypassed when enabling the MMU0.

TODO:
1. Adjust the logic for the MMU1 (for EDMA) in each DSP.
2. Evaluate if this programming needs to be done in the probe
   and release functions instead so that the DSP traffic always
   goes through the corresponding MMUs, irrespective of whether
   the actual MMU itself is enabled or not. It is currently
   not done since probe does not do any dynamic programming.

Signed-off-by: Suman Anna <s-anna@ti.com>