ARM: OMAP: remoteproc: request timers for DRA7xx DSP1 processor
The GPTimer #5 is chosen as the BIOS tick source for the DRA7xx
DSP1 processor, so add this timer data to be able to successfully
boot and run a DSP1 image configured with this timer. The timers
are optional, but are mandatory to support device management features
such as power management and watchdog support. The timer value also
needs to be unique from the ones used by other processors so that
they can be run simultaneously.

Signed-off-by: Suman Anna <s-anna@ti.com>