ARM: OMAP: remoteproc: assign timers for DRA7xx IPU processors
The GPTimers #3 and #11 are chosen as the BIOS tick source for the
DRA7xx IPU2 and IPU1 processors respectively, so add this timer data
to the DRA7xx remoteproc platform data. This is needed for successfully
booting an IPU image configured with these timers. The timers are
optional, but are mandatory to support device management features
such as power management and watchdog support.

The timer id for each processor needs to be unique so that other
processors can be run simultaneously. The IPUs are currently expected
to be running in SMP-mode, which requires only a single BIOS tick
timer per processor sub-system.

GPT11 is chosen for IPU2 based on its unique capabilities mask
previously, but the timer request logic no longer uses the capabilities
match logic. The GPT11 is chosen to maintain compatibility with the
firmware images.

Signed-off-by: Suman Anna <s-anna@ti.com>