video: da8xx-fb: prepare for adding hdmi encoder
This patch is in preparation for adding the hdmi encoder.
adding clock changes to support HDMI on am335x fix the divider
to 2 and change the value of the input clock.
It allows the addition of an hdmi= parameter in the lcdc device node
and a registration function so that the hdmi encoder can register with
the fbdev driver.
The hdmi= node is expected to point to a hdmi device node thus linking
the fbdev driver to a corresponding hdmi encoder driver.