video: da8xx-tda998x-hdmi: adding new driver
Adding a new HDMI encoder driver for the tda998x and the associated
calls into it.  This driver is based on the driver found in
gpu/drm/i2c/tda998x_drv.c but has been simplified for more minimal
functionality required.

Make the call into the set mode function, when a resolution needs to be
set.   This will set all the necessary timings into the hdmi encoder
device.

Signed-off-by: Darren Etheridge <detheridge@ti.com>