remoteproc: add api for retrieving the rproc platform id
A new api, rproc_get_platform_id, is added to allow remoteproc/rpmsg
clients to be able to retrieve a fixed id associated with a rproc
device. Each of the rproc's is given an ida index as well, but this
is dynamic depending on the order rproc_alloc is called, and also
depends on what remote processor devices are instantiated. The
platform device id is therefore chosen, with the assumption that
a remoteproc device is assigned a static platform device id.

The OMAP remoteproc devices do satisfy this requirement, since
static platform devices are needed anyway for declaring the
corresponding CMA pools.

Signed-off-by: Suman Anna <s-anna@ti.com>