rpmsg: rpc: lower the trace level in rppc_print_msg
The rppc_print_msg function is used for printing the
high-level decoded information about each rpmsg rpc
Tx and Rx packets. The trace levels of these traces
are lowered so that the traces are suppressed for
normal operation. This should improve the runtime
performance along with reducing the trace spew.

These traces can be enabled during debug, if needed,
by using the appropriate trace level.

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>