omapdss: hdmi pll: Add changes needed for OMAP5/DRA7x in terms of feats
Add a features struct to differentiate between the HDMI PLLs on OMAP4 and
OMAP5/DRA7x. The o5/dra7x PLL is more sensitive when ti comes to locking.
We need to ensure that the DCO freq isn't too low for lower pixel clocks.

Modify the PLL computation slightly to ensure the HDMI PLL locks for lower
frequencies. This will be replaced by a more complex computation which
makes sure all the PLL constraints are met.

Signed-off-by: Archit Taneja <archit@ti.com>