omapdss: hdmi4/hdmi5: set regulator voltage to required level
It's not sufficient to request the regulator. We need to set the
IO voltage as well when the min and max voltage values differ.

However, if the regulator on a platform doesn't permit the range
we requested for, or is a fixed voltage regulator, then just give a
warning that we didn't get voltage range we requested for.

Originally worked on by Somnath Mukherjee <somnath@ti.com>

Signed-off-by: Archit Taneja <archit@ti.com>