ARM: dts: AM4372: Add mailbox node
The mailbox DT node data has been added for AM33xx device, with
information added currently only for communicating with WkupM3
processor. The mailbox data is identical to that of the AM33xx
device. The functionality of WkupM3 processor mailbox is also
identical to that of the AM33xx device.

The usr_id value in the DT node reflects the value used for MPU
like the rest of the SoCs. The driver logic will be adjusted to
account for the WkupM3 usr_id within the code.

Signed-off-by: Suman Anna <s-anna@ti.com>