ARM: DRA7: hwmod_data: Add mailbox hwmod data
Add the hwmod data for the 13 instances of the system mailbox
IP in DRA7 SoC. The patch is needed for performing a soft-reset
while configuring the respective mailbox instance, otherwise is
a non-essential change for functionality. The modules are smart
idled on reset, and the IP module mode is hardware controlled.

NOTE: This patch requires the corresponding DT nodes to be
defined. This patch alone would cause a boot hang as the OMAP
boot code tries to enable and idle all the hwmods. The hwmod
data alone doesn't supply the module base addresses, and require
the regs definition from the DT nodes for properly configuring
the mailbox hwmods.

Signed-off-by: Suman Anna <s-anna@ti.com>