ARM: dts: DRA7: Add crossbar mappings for mailboxes
The MPU crossbar mappings are added for the mailboxes 5 and
6. Mailbox instance 1 has a default mapping. Mailboxes 5 and
6 are currently being used for signalling to the different
remote processors in the SoC.

The following are the mappings, the numbers are chosen based
on the current assignment in IPC 3.x, and needs to match the
interrupts field in the respective DT nodes. The crossbar
inputs are derived by assigning the MPU to the mailboxes'
user id 2.
   Mailbox 5: Crossbar input #251 mapped on MPU IRQ #136
   Mailbox 6: Crossbar input #255 mapped on MPU IRQ #141

The crossbar mapping in the kernel through the crossbar driver
allows independence from any required u-boot or pre-kernel
crossbar programming.

Signed-off-by: Suman Anna <s-anna@ti.com>