arm: dra7.dtsi: Add DSS DT information
Add DSS DT information. DSS on DRA7x is quite similar to OMAP5. The DSI blocks
are removed, but the PLLs are still kept and have given them different names.
RFBI is removed.

There are 3 DPI interfaces instead of one. The first one can receive it's data
from any of the 4 overlay managers. So the DT has an extra parameter called
channel for it. We could remove this if we dynamically allocate overlay
managers to encoders in the future. The rest have fixed overlay managers,
so they don't need anything.

The HDMI IP is the same as in OMAP5.

There are some nasty muxes from the control module which we need to configure
how the PLL outputs go back to DISPC. We need to figure out if they need a DT
representation for that.

Signed-off-by: Archit Taneja <archit@ti.com>