arm: dra7-evm.dts: Add LCD panel support
Add a node for pcf8575 gpio expander as a child of i2c0. One of the gpios from
this expander connects to the TLC gpio/backlight controller chip. The TLC chip
further powers on some pins of the DPI lcd panel.

The TLC chip and panel driver are clubbed for now. So a single TLC node will
suffice for now.

Signed-off-by: Archit Taneja <archit@ti.com>