omapdss: hdmi pll: Add feature for DRA7x to enable PLL via external mux
OMAP5 and DRA7x HDMI PLLs defer in terms of how the registers can be accessed.
On OMAP5, the PLL registers  are always accessed via the DSS domain. On DRA7x,
the PLLs can also be accessed via an ocp2scp interface, in this case, we don't
need to enable the DSS clock domain.

There is a control module register which selects the mode of access. Add a
feature parameter which configures PLL access via the DSS clock domain.

Signed-off-by: Archit Taneja <archit@ti.com>