arm: dra7-evm.dts: Add support for HDMI connector
Add the DT display node for HDMI connector on dra7x evm. The HDMI DDC lines are
connected to a demux. The demux is controlled via a mcasp2 gpio. The TPD level
shifter pins are controlled via a pcf8575 gpio expander.

The dts file just sees nodes for the io expander, the level shifter and the TPD
level shifter, the demux and the gpio to control the demux are hidden in the TPD
encoder driver.

Signed-off-by: Archit Taneja <archit@ti.com>