ARM: dts: AM4372: Fix mailbox node interrupt binding
The AM4372 DT uses 3-cells for interrupts, the existing mailbox
node is based on the AM33xx node that used a 1-cell interrupt
value. Correct this interrupt cell binding, so that the driver
can request the irq properly.

This fix is needed for proper functionality of mailbox on AM43x.

Signed-off-by: Suman Anna <s-anna@ti.com>