ARM: OMAP4+: Fix PRM device instance offsets for OMAP5 and DRA7
The PRM device instance offsets for OMAP5 and DRA7 are different
from that on OMAP4. While this was fixed in the
omap4_prminst_global_warm_sw_reset() function, it was fixed only
partially for the register read part. Fix the offsets in the rest
of the read/wriites done in the function.

Reported-by: Yan Liu <yan-liu@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>