hwspinlock/core: add common dt bindings and OF helpers
This patch adds the necessary documentation and OF helpers to
represent a hwlock device and use/request locks in a device-tree
build.

All the platform-specific hwlock driver implementations need the
number of locks and associated base id for registering the locks
present within the device with the driver core. The number of locks
is represented by 'hwlock-num-locks' property in DT bindings. A
property for base id is not needed in DT binding, as it can be
satisfied using a phandle + args specifier.

OF helpers have been added to the driver core to read the common
'hwlock-num-locks' property, and to request a specific lock using
the phandle + args specifier. The latter function is different
from the existing non-DT function, in that the specifier is
relative to the hwlock device instead of being a global lock id
registered with the core.

Signed-off-by: Suman Anna <s-anna@ti.com>